Colocate circuit-evolution, and give viz/flow its own NodeStyle
viz/flow/CustomNodes.tsx imported NodeStyle from this piece — a layer
reaching up into the thing it serves. It is the NAND node's own style
contract, so it belongs beside the component. viz/ now imports nothing
from a piece.
