+++ pages/blocks/circuit-evolution/figures/Circuit.tsx
