+++ pages/blocks/circuit-evolution/figures/CircuitDashboard.tsx
