+++ pages/blocks/circuit-evolution/figures/CircuitEvolutionSimulation.tsx
