+++ pages/blocks/circuit-evolution/figures/config.ts
