+++ pages/blocks/circuit-evolution/figures/data.ts
