+++ pages/blocks/circuit-evolution/figures/hooks.ts
