+++ pages/blocks/circuit-evolution/figures/index.ts
@@ -1,0 +1,1 @@
+export { default as CircuitEvolutionSimulation } from "./CircuitEvolutionSimulation";
