+++ pages/blocks/circuit-evolution/index.page.mdx
@@ -3,18 +3,10 @@ import { Accordion } from "@ui/controls";
 import { ArticleSection } from "@ui/layout";
 import { UnorderedList } from "@ui/prose";
 import MathFormula from "@ui/prose/MathFormula";
-import CircuitEvolutionSimulation from "../../../interactives/circuit-evolution/CircuitEvolutionSimulation";
+import { CircuitEvolutionSimulation } from "./figures";
+import meta from "./meta";
 
-export const metadata = {
-  title: "Circuit Evolution Simulator",
-  subtitle: "A basic model of how genetic circuits can evolve",
-  description: "An interactive circuit evolution simulator: mutate a network of NAND gates step by step, or select the fittest of each generation, and watch fitness climb.",
-  keywords: "circuit evolution, simulation, graph theory, probability",
-  ogImage: "https://www.newtinteractive.com/images/og/generations.png",
-  url: "https://www.newtinteractive.com/blocks/circuit-evolution",
-  published: "2024-11-26",
-  updated: "2024-12-02",
-};
+export const metadata = meta;
 
 export function InfoAccordion() {
   return (
