+++ viz/flow/index.ts
@@ -10,3 +10,4 @@ export { default as CircuitProteinNode } from "./custom-nodes/CircuitProteinNode
 export { default as CircuitPromoterNode } from "./custom-nodes/CircuitPromoterNode";
 export { default as LineNode } from "./custom-nodes/LineNode";
 export { default as LabelNode } from "./custom-nodes/LabelNode";
+export type { NodeStyle } from "./types";
