+++ viz/flow/types.ts
@@ -1,0 +1,5 @@
+export interface NodeStyle {
+  backgroundColor: string;
+  color: string;
+  borderColor?: string;
+}
